Moved /embed commands to a new feature.
